run-time bug fixed till ReceiveHealth

+++ GBServer/src/GBServer.cpp
@@ -87,7 +87,7 @@ namespace GenericBoson
 	{
 		WriteByteByByte(&pSi->m_writeBuffer, (int32_t)PacketType::Inventory);
 		WriteByteByByte(&pSi->m_writeBuffer, pSi->m_controllableCharacter.m_inventory.m_ID);
-		int16_t slotCount = (int16_t)pSi->m_controllableCharacter.m_inventory.GetTotalSlotCount();
+		int32_t slotCount = pSi->m_controllableCharacter.m_inventory.GetTotalSlotCount();
 		WriteByteByByte(&pSi->m_writeBuffer, slotCount);
 
 		for (auto& pSlot : pSi->m_controllableCharacter.m_inventory.m_slotVector)
